Add/Removed callback event handlers based on the setting.
